requirements.txt
# Core ML and NLP libraries
torch>=1.9.0
transformers>=4.20.0
accelerate>=0.26.0
datasets>=2.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Web framework
Flask>=2.0.0
Flask-CORS>=3.0.0

# Reddit API
praw>=7.6.0

# Data processing
requests>=2.28.0
beautifulsoup4>=4.11.0

# Production utilities
gunicorn>=20.1.0
python-dotenv>=0.19.0

# Logging and monitoring
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-flask>=1.2.0

# Development
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0